hack for the 32nd flag for now, new solution in the works.
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@10480 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/switch_ivr_bridge.c
@@ -529,7 +529,7 @@ static switch_status_t uuid_bridge_on_reset(switch_core_session_t *session)
 	switch_channel_clear_flag(channel, CF_TRANSFER);
 	switch_channel_clear_flag(channel, CF_ORIGINATING);
 
-	if (switch_channel_test_flag(channel, CF_MASTER)) {
+	if (switch_channel_test_flag(channel, (uint32_t)CF_MASTER)) {
 		switch_channel_set_state(channel, CS_SOFT_EXECUTE);
 	}
 
@@ -547,11 +547,11 @@ static switch_status_t uuid_bridge_on_soft_execute(switch_core_session_t *sessio
 
 	switch_channel_clear_flag(channel, CF_TRANSFER);
 
-	if (!switch_channel_test_flag(channel, CF_MASTER)) {
+	if (!switch_channel_test_flag(channel, (uint32_t)CF_MASTER)) {
 		return SWITCH_STATUS_SUCCESS;
 	}
 
-	switch_channel_clear_flag(channel, CF_MASTER);
+	switch_channel_clear_flag(channel, (uint32_t)CF_MASTER);
 
 	if ((other_uuid = switch_channel_get_variable(channel, SWITCH_UUID_BRIDGE)) && (other_session = switch_core_session_locate(other_uuid))) {
 		switch_channel_t *other_channel = switch_core_session_get_channel(other_session);
@@ -1035,8 +1035,8 @@ SWITCH_DECLARE(switch_status_t) switch_ivr_uuid_bridge(const char *originator_uu
 
 			switch_channel_clear_state_handler(originator_channel, NULL);
 			switch_channel_clear_state_handler(originatee_channel, NULL);
-			switch_channel_set_flag(originator_channel, CF_MASTER);
-			switch_channel_clear_flag(originatee_channel, CF_MASTER);
+			switch_channel_set_flag(originator_channel, (uint32_t)CF_MASTER);
+			switch_channel_clear_flag(originatee_channel, (uint32_t)CF_MASTER);
 			switch_channel_add_state_handler(originator_channel, &uuid_bridge_state_handlers);
 			switch_channel_add_state_handler(originatee_channel, &uuid_bridge_state_handlers);
 			switch_channel_set_variable(originator_channel, SWITCH_UUID_BRIDGE, switch_core_session_get_uuid(originatee_session));
